requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
langchain==0.1.12
langchain-google-genai==0.0.11
langchain-experimental==0.0.50
google-generativeai==0.4.1
numpy==1.26.4
pandas==2.2.1
PyPDF2==3.0.1
PyMuPDF==1.25.5
Pillow==10.2.0
scikit-learn==1.4.0
graphviz==0.20.1